Fixed issue with 3D objects loading !

+++ Identity/Engine/src/3DLoader/ObjectElements/Mesh.cpp
@@ -2,7 +2,7 @@
 
 #include <3DLoader/ObjectElements/Mesh.h>
 
-Engine::ObjectElements::Mesh::Mesh(std::vector<Engine::Geometry::Vertex>& p_vertices, std::vector<uint32_t>& p_indices) :
+Engine::ObjectElements::Mesh::Mesh(std::vector<Engine::Geometry::Vertex>& p_vertices, std::vector<unsigned short>& p_indices) :
     m_vertices { p_vertices }, m_indices { p_indices } {}
 
 Engine::ObjectElements::Mesh::Mesh(const Mesh& p_other):
